app/requirements.txt
#Pyhton dependencies

alembic==1.11.1
anyio==3.7.0
astroid==2.15.5
asyncpg==0.27.0
boto3==1.26.143
botocore==1.29.143
click==8.1.3
colorama==0.4.6
dill==0.3.6
exceptiongroup==1.1.1
fastapi==0.95.2
greenlet==2.0.2
h11==0.14.0
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
isort==5.12.0
jmespath==1.0.1
lazy-object-proxy==1.9.0
Mako==1.2.4
netcat
python-dotenv
MarkupSafe==2.1.2
mccabe==0.7.0
pip==21.1.1
platformdirs==3.6.0
psycopg2==2.9.6
pydantic==1.10.8
pylint==2.17.4
python-dateutil==2.8.2
s3transfer==0.6.1
setuptools==56.0.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.15
starlette==0.27.0
tomli==2.0.1
tomlkit==0.11.8
typing-extensions==4.6.2
urllib3==1.26.16
uvicorn==0.22.0
wrapt==1.15.0
zipp==3.15.0
